src/main/java/com/monalloyd/katas/findtheuniquenumber/UniqueNumberFinder.java
package com.monalloyd.katas.findtheuniquenumber;

import java.util.Arrays;
import java.util.Map;
import java.util.Optional;
import java.util.stream.Collectors;

public class UniqueNumberFinder {
    public Optional<Double> findUnique(double[] arr) {
        return Arrays.stream(arr)
                .boxed()
                .collect(Collectors.groupingBy(num -> num, Collectors.counting()))
                .entrySet().stream()
                .filter(entry -> entry.getValue() == 1)
                .findFirst()
                .map(Map.Entry::getKey);
    }
}
